Moving Constants class to util package
it's not really a model, an entity
